+++ Sources/SwiftDate/Commons.swift
@@ -1,28 +1,53 @@
+// SwiftDate
+// Manage Date/Time & Timezone in Swift
 //
-//	SwiftDate, Full featured Swift date library for parsing, validating, manipulating, and formatting dates and timezones.
-//	Created by:				Daniele Margutti
-//	Main contributors:		Jeroen Houtzager
+// Created by: Daniele Margutti
+// Email: <hello@danielemargutti.com>
+// Web: <http://www.danielemargutti.com>
 //
-//
-//	Permission is hereby granted, free of charge, to any person obtaining a copy
-//	of this software and associated documentation files (the "Software"), to deal
-//	in the Software without restriction, including without limitation the rights
-//	to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-//	copies of the Software, and to permit persons to whom the Software is
-//	furnished to do so, subject to the following conditions:
-//
-//	The above copyright notice and this permission notice shall be included in
-//	all copies or substantial portions of the Software.
-//
-//	THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-//	IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-//	FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-//	AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-//	LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-//	OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-//	THE SOFTWARE.
+// Licensed under MIT License.
 
 import Foundation
+
+internal extension Calendar.Component {
+	
+	/// Return the localized identifier of a calendar component
+	///
+	/// - parameter unit:  unit
+	/// - parameter value: value
+	///
+	/// - returns: return the plural or singular form of the time unit used to compose a valid identifier for search a localized
+	///   string in resource bundle
+	internal func localizedKey(forValue value: Int) -> String {
+		let locKey = self.localizedKey
+		let absValue = abs(value)
+		switch absValue {
+		case 0: // zero difference for this unit
+			return "0\(locKey)"
+		case 1: // one unit of difference
+			return locKey
+		default: // more than 1 unit of difference
+			return "\(locKey)\(locKey)"
+		}
+	}
+	
+	
+	internal var localizedKey: String {
+		switch self {
+		case .year:			return "y"
+		case .month:		return "m"
+		case .weekOfYear:	return "w"
+		case .day:			return "d"
+		case .hour:			return "h"
+		case .minute:		return "M"
+		case .second:		return "s"
+		default:
+			return ""
+		}
+	}
+	
+}
+
 
 /// Time interval reference
 ///
@@ -199,8 +224,8 @@ public struct ComponentsFormatterOptions {
 	}
 }
 
-private let SECONDS_IN_MINUTE: TimeInterval = 60
-private let SECONDS_IN_HOUR: TimeInterval = SECONDS_IN_MINUTE * 60
-private let SECONDS_IN_DAY: TimeInterval = SECONDS_IN_HOUR * 24
-private let SECONDS_IN_WEEK: TimeInterval = SECONDS_IN_DAY * 7
+internal let SECONDS_IN_MINUTE: TimeInterval = 60
+internal let SECONDS_IN_HOUR: TimeInterval = SECONDS_IN_MINUTE * 60
+internal let SECONDS_IN_DAY: TimeInterval = SECONDS_IN_HOUR * 24
+internal let SECONDS_IN_WEEK: TimeInterval = SECONDS_IN_DAY * 7