+++ Sources/SwiftDate/Date+Compare.swift
@@ -1,30 +1,13 @@
+// SwiftDate
+// Manage Date/Time & Timezone in Swift
 //
-//	SwiftDate, Full featured Swift date library for parsing, validating, manipulating, and formatting dates and timezones.
-//	Created by:				Daniele Margutti
-//	Main contributors:		Jeroen Houtzager
+// Created by: Daniele Margutti
+// Email: <hello@danielemargutti.com>
+// Web: <http://www.danielemargutti.com>
 //
-//
-//	Permission is hereby granted, free of charge, to any person obtaining a copy
-//	of this software and associated documentation files (the "Software"), to deal
-//	in the Software without restriction, including without limitation the rights
-//	to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-//	copies of the Software, and to permit persons to whom the Software is
-//	furnished to do so, subject to the following conditions:
-//
-//	The above copyright notice and this permission notice shall be included in
-//	all copies or substantial portions of the Software.
-//
-//	THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-//	IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-//	FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-//	AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-//	LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-//	OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-//	THE SOFTWARE.
+// Licensed under MIT License.
 
 import Foundation
-
-// MARK: - Date Extensions to Compare
 
 public extension Date {
 	
@@ -92,5 +75,21 @@ public extension Date {
 		let result = self.compare(to: date, granularity: granularity)
 		return (orEqual ? (result == .orderedSame || result == .orderedDescending) : result == .orderedDescending)
 	}
+    
+    
+    /// Compares whether a date is between two dates based on their components down to a given unit
+    /// granularity.
+    /// Both passed `Date` objects are expressed in passed region before doing the comparison.
+    ///
+    /// - parameter date:        first date as lower bound
+    /// - parameter date2:       second date as upper bound
+    /// - parameter orEqual:     `true` to also check for equality on date and date2
+    /// - parameter region:      region in which both dates will be expressed (if nil, defaultRegion will be used instead)
+    /// - parameter granularity: The smallest unit that must, along with all larger units, be greater
+    ///
+    /// - returns: `true` if the unit is in the middle of two dates, otherwise `false`
+    public func isBetween(date: Date, and date2: Date, orEqual: Bool = false, in region: Region? = nil, granularity: Calendar.Component = .nanosecond) -> Bool{
+        return self.isAfter(date: date, orEqual: orEqual, in: region, granularity: granularity) && self.isBefore(date: date2, orEqual: orEqual, in: region, granularity: granularity)
+    }
 	
 }